Convert changes for TypeScript, terribly fun
