+++ db/index.ts
@@ -1,1 +1,2 @@
 export * from './models/splash.model';
+export * from './models/info.model';
