+++ db/models/info.model.ts
@@ -1,0 +1,6 @@
+import db from './../mongoose';
+import InfoSchema from './../schema/info/info.schema';
+
+const InfoModel = db.model('info', InfoSchema);
+
+export { InfoModel };
