--- db/models/project.model.js
@@ -1,6 +1,0 @@
-const db = require('./../mongoose');
-const ProjectSchema = require('./../schema/project/project.schema');
-
-const ProjectModel = db.model('project', ProjectSchema);
-
-module.exports = ProjectModel;
